requirements.txt
fastapi
uvicorn
python-multipart
jinja2
opencv-python
torch
transformers
pillow
numpy<2
gTTS
deep-translator
ultralytics
